+++ config/splat.dra.yaml
@@ -70,7 +70,6 @@ segments:
       - [0x5D7C0, c, 5D7C0]
       - [0x5D874, c, 5D874]
       - [0x73AAC, c, 73AAC]
-      - [0x73D7C, c, 73D7C]
       - [0x75DA0, c, 75DA0]
       - [0x75F54, c, 75F54]
       - [0x953A0, c, 953A0]
